+++ core/src/bms/player/beatoraja/input/KeyCommand.java
@@ -15,5 +15,5 @@ public enum KeyCommand {
 	AUTOPLAY_FOLDER,
 	OPEN_IR,
 	OPEN_SKIN_CONFIGURATION,
-	TOGGLE_TRAINER,
+    TOGGLE_MOD_MENU,
 }
